Début de l'implémentation de la grammaire

+++ html/constante.cc
@@ -1,6 +1,6 @@
 #include "constante.hh"
 
-Constante::Constante(double valeur)
+Constante::Constante(int valeur)
     : _valeur(valeur) {
 
 }
@@ -9,6 +9,6 @@ std::string Constante::to_html(const Contexte & contexte) const {
     return std::to_string(_valeur);
 }
 
-double Constante::valeur() const {
+int Constante::valeur() const {
     return _valeur;
 }
